# assignment2/api_service/server/python-flask/openapi_server/controllers/default_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server.models.auth_response import AuthResponse  # noqa: E501
from openapi_server.models.error_response import ErrorResponse  # noqa: E501
from openapi_server.models.get_greeting200_response import GetGreeting200Response  # noqa: E501
from openapi_server.models.list_products200_response import ListProducts200Response  # noqa: E501
from openapi_server.models.order_create import OrderCreate  # noqa: E501
from openapi_server.models.success_response import SuccessResponse  # noqa: E501
from openapi_server.models.user_login import UserLogin  # noqa: E501
from openapi_server.models.user_registration import UserRegistration  # noqa: E501
from openapi_server.models.user_update import UserUpdate  # noqa: E501
from openapi_server import util


def cancel_order(id):  # noqa: E501
    """Cancel order

    Cancel a specific order by ID for the authenticated user # noqa: E501

    :param id: 
    :type id: int

    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def delete_current_user():  # noqa: E501
    """Delete current user account

    Deactivate the authenticated user&#39;s account # noqa: E501


    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_current_user():  # noqa: E501
    """Get current user profile

    Retrieve the authenticated user&#39;s profile information # noqa: E501


    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_greeting():  # noqa: E501
    """Get welcome message

    Returns a welcome message for the SUSTech Merch Store # noqa: E501


    :rtype: Union[GetGreeting200Response, Tuple[GetGreeting200Response, int], Tuple[GetGreeting200Response, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_order(id):  # noqa: E501
    """Get order by ID

    Retrieve a specific order by ID for the authenticated user # noqa: E501

    :param id: 
    :type id: int

    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_product(id):  # noqa: E501
    """Get product by ID

    Retrieve detailed information about a specific product # noqa: E501

    :param id: 
    :type id: int

    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_user_orders():  # noqa: E501
    """Get user&#39;s orders

    Retrieve all orders for the authenticated user # noqa: E501


    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def list_products():  # noqa: E501
    """List all products

    Retrieve a list of all available products # noqa: E501


    :rtype: Union[ListProducts200Response, Tuple[ListProducts200Response, int], Tuple[ListProducts200Response, int, Dict[str, str]]
    """
    return 'do some magic!'


def login_user(user_login):  # noqa: E501
    """User login

    Authenticate user and return JWT token # noqa: E501

    :param user_login: 
    :type user_login: dict | bytes

    :rtype: Union[AuthResponse, Tuple[AuthResponse, int], Tuple[AuthResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        user_login = UserLogin.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def place_order(order_create):  # noqa: E501
    """Place a new order

    Create a new order for the authenticated user # noqa: E501

    :param order_create: 
    :type order_create: dict | bytes

    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        order_create = OrderCreate.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def register_user(user_registration):  # noqa: E501
    """Register a new user

    Create a new user account # noqa: E501

    :param user_registration: 
    :type user_registration: dict | bytes

    :rtype: Union[AuthResponse, Tuple[AuthResponse, int], Tuple[AuthResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        user_registration = UserRegistration.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def update_current_user(user_update):  # noqa: E501
    """Update current user profile

    Update the authenticated user&#39;s profile information # noqa: E501

    :param user_update: 
    :type user_update: dict | bytes

    :rtype: Union[SuccessResponse, Tuple[SuccessResponse, int], Tuple[SuccessResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        user_update = UserUpdate.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'