# assignment2/api_service/server/python-flask/openapi_server/models/error_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ErrorResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, success=None, error=None, code=None):  # noqa: E501
        """ErrorResponse - a model defined in OpenAPI

        :param success: The success of this ErrorResponse.  # noqa: E501
        :type success: bool
        :param error: The error of this ErrorResponse.  # noqa: E501
        :type error: str
        :param code: The code of this ErrorResponse.  # noqa: E501
        :type code: int
        """
        self.openapi_types = {
            'success': bool,
            'error': str,
            'code': int
        }

        self.attribute_map = {
            'success': 'success',
            'error': 'error',
            'code': 'code'
        }

        self._success = success
        self._error = error
        self._code = code

    @classmethod
    def from_dict(cls, dikt) -> 'ErrorResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ErrorResponse of this ErrorResponse.  # noqa: E501
        :rtype: ErrorResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def success(self) -> bool:
        """Gets the success of this ErrorResponse.


        :return: The success of this ErrorResponse.
        :rtype: bool
        """
        return self._success

    @success.setter
    def success(self, success: bool):
        """Sets the success of this ErrorResponse.


        :param success: The success of this ErrorResponse.
        :type success: bool
        """

        self._success = success

    @property
    def error(self) -> str:
        """Gets the error of this ErrorResponse.


        :return: The error of this ErrorResponse.
        :rtype: str
        """
        return self._error

    @error.setter
    def error(self, error: str):
        """Sets the error of this ErrorResponse.


        :param error: The error of this ErrorResponse.
        :type error: str
        """

        self._error = error

    @property
    def code(self) -> int:
        """Gets the code of this ErrorResponse.


        :return: The code of this ErrorResponse.
        :rtype: int
        """
        return self._code

    @code.setter
    def code(self, code: int):
        """Sets the code of this ErrorResponse.


        :param code: The code of this ErrorResponse.
        :type code: int
        """

        self._code = code