# assignment2/api_service/server/python-flask/openapi_server/models/order_create.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class OrderCreate(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, product_id=None, quantity=None):  # noqa: E501
        """OrderCreate - a model defined in OpenAPI

        :param product_id: The product_id of this OrderCreate.  # noqa: E501
        :type product_id: int
        :param quantity: The quantity of this OrderCreate.  # noqa: E501
        :type quantity: int
        """
        self.openapi_types = {
            'product_id': int,
            'quantity': int
        }

        self.attribute_map = {
            'product_id': 'product_id',
            'quantity': 'quantity'
        }

        self._product_id = product_id
        self._quantity = quantity

    @classmethod
    def from_dict(cls, dikt) -> 'OrderCreate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The OrderCreate of this OrderCreate.  # noqa: E501
        :rtype: OrderCreate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def product_id(self) -> int:
        """Gets the product_id of this OrderCreate.


        :return: The product_id of this OrderCreate.
        :rtype: int
        """
        return self._product_id

    @product_id.setter
    def product_id(self, product_id: int):
        """Sets the product_id of this OrderCreate.


        :param product_id: The product_id of this OrderCreate.
        :type product_id: int
        """
        if product_id is None:
            raise ValueError("Invalid value for `product_id`, must not be `None`")  # noqa: E501

        self._product_id = product_id

    @property
    def quantity(self) -> int:
        """Gets the quantity of this OrderCreate.


        :return: The quantity of this OrderCreate.
        :rtype: int
        """
        return self._quantity

    @quantity.setter
    def quantity(self, quantity: int):
        """Sets the quantity of this OrderCreate.


        :param quantity: The quantity of this OrderCreate.
        :type quantity: int
        """
        if quantity is None:
            raise ValueError("Invalid value for `quantity`, must not be `None`")  # noqa: E501
        if quantity is not None and quantity > 3:  # noqa: E501
            raise ValueError("Invalid value for `quantity`, must be a value less than or equal to `3`")  # noqa: E501
        if quantity is not None and quantity < 1:  # noqa: E501
            raise ValueError("Invalid value for `quantity`, must be a value greater than or equal to `1`")  # noqa: E501

        self._quantity = quantity